requirements.txt
annotated-types==0.7.0
anyio==4.8.0
asgiref==3.8.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cloudinary==1.42.2
cryptography==44.0.2
defusedxml==0.7.1
distro==1.9.0
dj-database-url==0.5.0
dj3-cloudinary-storage==0.0.6
Django==4.2.19
django-allauth==0.57.2
django-cors-headers==4.7.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
gunicorn==20.1.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jiter==0.8.2
lxml==5.3.1
oauthlib==3.2.2
openai==1.65.2
psycopg2==2.9.10
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
PyJWT==2.9.0
PyPDF2==3.0.1
python-docx==1.1.2
python3-openid==3.2.0
requests==2.32.3
requests-oauthlib==2.0.0
setuptools==75.8.0
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==1.26.20
whitenoise==5.3.0
